+++ tests/Attributes/KeyTest.php
@@ -31,11 +31,17 @@ class KeyTest extends TestCase
         $this->assertSame('layout_2f1c419c', $key);
     }
 
-    public function testKeyUniqueness()
+    public function testValidate()
     {
         $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage('The key [layout_2f1c419c] is not unique.');
 
         Key::generate('layout', 'block_image');
+
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage('The key prefix must be either field, group or layout.');
+
+        Key::validate('image');
     }
 
     public function testSanitize()